+++ inst/shiny/modules/dispersion.R
@@ -18,13 +18,15 @@ dispersionUiInput <- function(){
       condition = "input.dispersion_object == 'corpus'",
       selectInput("dispersion_corpus", "corpus", corpus()[["corpus"]])
     ),
+    radioButtons("dispersion_freq", "frequencies", choices = list("yes", "no"), selected = "no", inline = TRUE),
+    radioButtons("dispersion_cqp", "CQP", choices = list("yes", "no"), selected = "no", inline = TRUE),
     conditionalPanel(
       condition = "input.dispersion_object == 'partition'",
       selectInput("dispersion_partition", "partition", choices = character())
     ),
     textInput("dispersion_query", "query", value = "Suche"),
     selectInput(
-      "dispersion_p_Attribute", "p_attribute",
+      "dispersion_p_attribute", "p_attribute",
       choices = p_attributes(corpus()[1, "corpus"])
     ),
     # radioButtons("dispersion_dims", "number of s_attributes", choices=c("1", "2"), selected="1", inline=TRUE),
@@ -75,44 +77,50 @@ dispersionServer <- function(input, output, session){
     input$dispersion_go,
     isolate({
       
-      if (input$dispersion_object == "partition"){
-        object <- values$partitions[[input$dispersion_partition]]
-      } else if (input$dispersion_object == "corpus"){
-        object <- input$dispersion_corpus
-      }
-      
-      tab <- as.data.frame(dispersion(
-        object,
-        query = input$dispersion_query,
+      obj <- if (input$dispersion_object == "partition")
+        values$partitions[[input$dispersion_partition]]
+      else 
+        input$dispersion_corpus
+
+      tab <- dispersion(
+        obj,
+        query = rectifySpecialChars(input$dispersion_query),
         s_attribute = input$dispersion_s_attribute_1,
-        p_attribute = input$dispersion_p_attribute
-      ))
+        p_attribute = input$dispersion_p_attribute,
+        cqp = if (input$dispersion_cqp == "yes") TRUE else FALSE,
+        freq = if (input$dispersion_freq == "yes") TRUE else FALSE
+      )
       
-      # tab[["freq"]] <- round(tab[["freq"]], 7)
-      
-      
+      if ("freq" %in% colnames(tab)) tab[, "freq" := round(tab[["freq"]] * 100000, 7)]
+
       if (input$dispersion_ts == "yes"){
-        dates4zoo <- tab[[input$dispersion_s_attribute_1]]
-        tab4zoo <- data.table::copy(tab)
-        tab4zoo[[input$dispersion_s_attribute_1]] <- NULL
-        zooObject <- zoo::zoo(as.matrix(tab4zoo), order.by = as.Date(dates4zoo))
+        dates <- tab[[input$dispersion_s_attribute_1]]
+        dt <- data.table::copy(tab)
+        print(is(dt))
+        print(head(dt))
+        dt[, eval(input$dispersion_s_attribute_1) := NULL]
+        if ("query" %in% colnames(dt)) dt[, "query" := NULL]
+        print(head(dt))
+        zoo_obj <- zoo::zoo(as.matrix(dt), order.by = as.Date(dates))
+        print(head(zoo_obj))
+        
         if (input$dispersion_ts_aggregation != "none"){
-          zooObject <- switch(
+          zoo_obj <- switch(
             input$dispersion_ts_aggregation,
-            month = aggregate(zooObject, zoo::as.Date(zoo::as.yearmon(zoo::index(zooObject)))),
-            quarter = aggregate(zooObject, zoo::as.Date(zoo::as.yearqtr(zoo::index(zooObject)))),
-            year = aggregate(zooObject, as.Date(paste(gsub("^(\\d{4}).*?$", "\\1", zoo::index(zooObject)), "-01-01", sep="")))
+            month = aggregate(zoo_obj, zoo::as.Date(zoo::as.yearmon(zoo::index(zoo_obj)))),
+            quarter = aggregate(zoo_obj, zoo::as.Date(zoo::as.yearqtr(zoo::index(zoo_obj)))),
+            year = aggregate(zoo_obj, as.Date(paste(gsub("^(\\d{4}).*?$", "\\1", zoo::index(zoo_obj)), "-01-01", sep="")))
           )
           # rework data.frame
-          tab <- data.frame(date = zoo::index(zooObject), zooObject)
+          tab <- data.frame(date = zoo::index(zoo_obj), zoo_obj)
           colnames(tab)[1] <- input$dispersion_s_attribute_1
           rownames(tab) <- NULL
-          output$dispersion_plot <- renderPlot(zoo::plot.zoo(zooObject, main=""))
+          output$dispersion_plot <- renderPlot(zoo::plot.zoo(zoo_obj, main = ""))
         }
       } else {
         output$dispersion_plot <- renderPlot(
           barplot(
-            height = tab[["count"]],
+            height = if (input$dispersion_freq == "yes") tab[["freq"]] else tab[["count"]],
             names.arg = tab[[input$dispersion_s_attribute_1]]
           )
         )